chore: prepare for tauri v2
