+++ backend/tauri/src/core/tray/mod.rs
@@ -4,10 +4,15 @@ use crate::{
     utils::{help, resolve},
 };
 use anyhow::Result;
+use parking_lot::Mutex;
 use rust_i18n::t;
 use tauri::{
-    AppHandle, CustomMenuItem, SystemTrayEvent, SystemTrayMenu, SystemTrayMenuItem,
-    SystemTraySubmenu,
+    menu::{
+        CheckMenuItemBuilder, Menu, MenuBuilder, MenuEvent, MenuItemBuilder, PredefinedMenuItem,
+        SubmenuBuilder,
+    },
+    tray::{MouseButton, TrayIcon, TrayIconBuilder, TrayIconEvent},
+    App, AppHandle, Manager,
 };
 use tracing_attributes::instrument;
 
@@ -18,9 +23,14 @@ use self::proxies::SystemTrayMenuProxiesExt;
 mod utils;
 pub struct Tray {}
 
+pub struct TrayState {
+    pub tray_icon: TrayIcon,
+    pub menu: Mutex<Menu>,
+}
+
 impl Tray {
-    #[instrument(skip(_app_handle))]
-    pub fn tray_menu(_app_handle: &AppHandle) -> SystemTrayMenu {
+    #[instrument(skip(app_handle))]
+    pub fn tray_menu(app_handle: &AppHandle) -> Result<Menu> {
         let version = env!("NYANPASU_VERSION");
         let core = {
             *Config::verge()
@@ -29,66 +39,145 @@ impl Tray {
                 .as_ref()
                 .unwrap_or(&ClashCore::default())
         };
-        let mut menu = SystemTrayMenu::new()
-            .add_item(CustomMenuItem::new("open_window", t!("tray.dashboard")))
-            .setup_proxies() // Setup the proxies menu
-            .add_native_item(SystemTrayMenuItem::Separator)
-            .add_item(CustomMenuItem::new("rule_mode", t!("tray.rule_mode")))
-            .add_item(CustomMenuItem::new("global_mode", t!("tray.global_mode")))
-            .add_item(CustomMenuItem::new("direct_mode", t!("tray.direct_mode")));
+        let mut menu = MenuBuilder::new(app_handle)
+            .item(
+                &MenuItemBuilder::new(t!("tray.dashboard"))
+                    .id("open_window")
+                    .build(app_handle)?,
+            )
+            // .setup_proxies() // Setup the proxies menu
+            .item(&PredefinedMenuItem::separator(app_handle)?)
+            .item(
+                &CheckMenuItemBuilder::new(t!("tray.rule_mode"))
+                    .id("rule_mode")
+                    .build(app_handle)?,
+            )
+            .item(
+                &CheckMenuItemBuilder::new(t!("tray.global_mode"))
+                    .id("global_mode")
+                    .build(app_handle)?,
+            )
+            .item(
+                &CheckMenuItemBuilder::new(t!("tray.direct_mode"))
+                    .id("direct_mode")
+                    .build(app_handle)?,
+            );
         if core == ClashCore::ClashPremium {
-            menu = menu.add_item(CustomMenuItem::new("script_mode", t!("tray.script_mode")))
+            menu = menu.item(
+                &CheckMenuItemBuilder::new(t!("tray.script_mode"))
+                    .id("script_mode")
+                    .build(app_handle)?,
+            );
         }
-        menu.add_native_item(SystemTrayMenuItem::Separator)
-            .add_item(CustomMenuItem::new("system_proxy", t!("tray.system_proxy")))
-            .add_item(CustomMenuItem::new("tun_mode", t!("tray.tun_mode")))
-            .add_item(CustomMenuItem::new("copy_env_sh", t!("tray.copy_env.sh")))
-            .add_item(CustomMenuItem::new("copy_env_cmd", t!("tray.copy_env.cmd")))
-            .add_item(CustomMenuItem::new("copy_env_ps", t!("tray.copy_env.ps")))
-            .add_submenu(SystemTraySubmenu::new(
-                t!("tray.open_dir.menu"),
-                SystemTrayMenu::new()
-                    .add_item(CustomMenuItem::new(
-                        "open_app_config_dir",
-                        t!("tray.open_dir.app_config_dir"),
-                    ))
-                    .add_item(CustomMenuItem::new(
-                        "open_app_data_dir",
-                        t!("tray.open_dir.app_data_dir"),
-                    ))
-                    .add_item(CustomMenuItem::new(
-                        "open_core_dir",
-                        t!("tray.open_dir.core_dir"),
-                    ))
-                    .add_item(CustomMenuItem::new(
-                        "open_logs_dir",
-                        t!("tray.open_dir.log_dir"),
-                    )),
-            ))
-            .add_submenu(SystemTraySubmenu::new(
-                t!("tray.more.menu"),
-                SystemTrayMenu::new()
-                    .add_item(CustomMenuItem::new(
-                        "restart_clash",
-                        t!("tray.more.restart_clash"),
-                    ))
-                    .add_item(CustomMenuItem::new(
-                        "restart_app",
-                        t!("tray.more.restart_app"),
-                    ))
-                    .add_item(
-                        CustomMenuItem::new("app_version", format!("Version {version}")).disabled(),
-                    ),
-            ))
-            .add_native_item(SystemTrayMenuItem::Separator)
-            .add_item(CustomMenuItem::new("quit", t!("tray.quit")).accelerator("CmdOrControl+Q"))
+        menu.item(&PredefinedMenuItem::separator(app_handle)?)
+            .item(
+                &CheckMenuItemBuilder::new(t!("tray.system_proxy"))
+                    .id("system_proxy")
+                    .build(app_handle)?,
+            )
+            .item(
+                &CheckMenuItemBuilder::new(t!("tray.tun_mode"))
+                    .id("tun_mode")
+                    .build(app_handle)?,
+            )
+            .item(
+                &MenuItemBuilder::new(t!("tray.copy_env_sh"))
+                    .id("copy_env_sh")
+                    .build(app_handle)?,
+            )
+            .item(
+                &MenuItemBuilder::new(t!("tray.copy_env_cmd"))
+                    .id("copy_env_cmd")
+                    .build(app_handle)?,
+            )
+            .item(
+                &MenuItemBuilder::new(t!("tray.copy_env_ps"))
+                    .id("copy_env_ps")
+                    .build(app_handle)?,
+            )
+            .item(
+                &SubmenuBuilder::new(app_handle, t!("tray.open_dir.menu"))
+                    .item(
+                        &MenuItemBuilder::new(t!("tray.open_dir.app_config_dir"))
+                            .id("open_app_config_dir")
+                            .build(app_handle)?,
+                    )
+                    .item(
+                        &MenuItemBuilder::new(t!("tray.open_dir.app_data_dir"))
+                            .id("open_app_data_dir")
+                            .build(app_handle)?,
+                    )
+                    .item(
+                        &MenuItemBuilder::new(t!("tray.open_dir.core_dir"))
+                            .id("open_core_dir")
+                            .build(app_handle)?,
+                    )
+                    .item(
+                        &MenuItemBuilder::new(t!("tray.open_dir.log_dir"))
+                            .id("open_logs_dir")
+                            .build(app_handle)?,
+                    )
+                    .build()?,
+            )
+            .item(
+                &SubmenuBuilder::new(app_handle, t!("tray.more.menu"))
+                    .item(
+                        &MenuItemBuilder::new(t!("tray.more.restart_clash"))
+                            .id("restart_clash")
+                            .build(app_handle)?,
+                    )
+                    .item(
+                        &MenuItemBuilder::new(t!("tray.more.restart_app"))
+                            .id("restart_app")
+                            .build(app_handle)?,
+                    )
+                    .item(
+                        &MenuItemBuilder::new(format!("Version {}", version))
+                            .id("app_version")
+                            .enabled(false)
+                            .build(app_handle)?,
+                    )
+                    .build()?,
+            )
+            .item(&PredefinedMenuItem::separator(app_handle)?)
+            .item(
+                &MenuItemBuilder::new(t!("tray.quit"))
+                    .id("quit")
+                    .accelerator("CmdOrControl+Q")
+                    .build(app_handle)?,
+            );
+        Ok(menu.build()?)
+    }
+
+    pub fn setup_tray(app: &mut App) -> Result<()> {
+        let menu = Tray::tray_menu(app.handle())?;
+        let tray_icon = TrayIconBuilder::new()
+            .icon(tauri::image::Image::from_bytes(&icon::get_icon(
+                &icon::TrayIcon::Normal,
+            ))?)
+            .menu(&menu)
+            .on_menu_event(|app, event| {
+                Tray::on_menu_item_event(app, event);
+            })
+            .on_tray_icon_event(|tray_icon, event| {
+                Tray::on_system_tray_event(tray_icon, event);
+            })
+            .build(app)?;
+        app.manage::<TrayState>(TrayState {
+            tray_icon,
+            menu: Mutex::new(menu),
+        });
+        Ok(())
     }
 
     #[instrument(skip(app_handle))]
     pub fn update_systray(app_handle: &AppHandle) -> Result<()> {
-        app_handle
-            .tray_handle()
-            .set_menu(Tray::tray_menu(app_handle))?;
+        let tray_state = app_handle.state::<TrayState>();
+        let menu = Tray::tray_menu(app_handle)?;
+        tray_state.tray_icon.set_menu(Some(menu.clone()))?;
+        {
+            *tray_state.menu.lock() = menu;
+        }
         Tray::update_part(app_handle)?;
         Ok(())
     }
@@ -103,8 +192,8 @@ impl Tray {
                 .as_ref()
                 .unwrap_or(&ClashCore::default())
         };
-        let tray = app_handle.tray_handle();
-
+        let tray = app_handle.state::<TrayState>();
+        let menu = tray.menu.lock();
         #[cfg(target_os = "linux")]
         {
             let _ = tray.get_item("rule_mode").set_title(t!("tray.rule_mode"));
@@ -145,11 +234,19 @@ impl Tray {
         }
         #[cfg(not(target_os = "linux"))]
         {
-            let _ = tray.get_item("rule_mode").set_selected(mode == "rule");
-            let _ = tray.get_item("global_mode").set_selected(mode == "global");
-            let _ = tray.get_item("direct_mode").set_selected(mode == "direct");
+            let _ = menu
+                .get("rule_mode")
+                .and_then(|item| item.as_check_menuitem()?.set_checked(mode == "rule").ok());
+            let _ = menu
+                .get("global_mode")
+                .and_then(|item| item.as_check_menuitem()?.set_checked(mode == "global").ok());
+            let _ = menu
+                .get("direct_mode")
+                .and_then(|item| item.as_check_menuitem()?.set_checked(mode == "direct").ok());
             if core == ClashCore::ClashPremium {
-                let _ = tray.get_item("script_mode").set_selected(mode == "script");
+                let _ = menu
+                    .get("script_mode")
+                    .and_then(|item| item.as_check_menuitem()?.set_checked(mode == "script").ok());
             }
         }
 
@@ -174,7 +271,9 @@ impl Tray {
                 TrayIcon::Normal
             };
             let icon = icon::get_icon(&mode);
-            let _ = tray.set_icon(tauri::Icon::Raw(icon));
+            let _ = tray
+                .tray_icon
+                .set_icon(Some(tauri::image::Image::from_bytes(&icon)?));
         }
 
         #[cfg(target_os = "linux")]
@@ -206,8 +305,12 @@ impl Tray {
 
         #[cfg(not(target_os = "linux"))]
         {
-            let _ = tray.get_item("system_proxy").set_selected(system_proxy);
-            let _ = tray.get_item("tun_mode").set_selected(tun_mode);
+            let _ = menu
+                .get("system_proxy")
+                .and_then(|item| item.as_check_menuitem()?.set_checked(system_proxy).ok());
+            let _ = menu
+                .get("tun_mode")
+                .and_then(|item| item.as_check_menuitem()?.set_checked(tun_mode).ok());
         }
 
         #[cfg(not(target_os = "linux"))]
@@ -219,51 +322,54 @@ impl Tray {
                 map
             };
 
-            let _ = tray.set_tooltip(&format!(
+            let _ = tray.tray_icon.set_tooltip(Some(&format!(
                 "{}: {}\n{}: {}",
                 t!("tray.system_proxy"),
                 switch_map[&system_proxy],
                 t!("tray.tun_mode"),
                 switch_map[&tun_mode]
-            ));
+            )));
         }
 
         Ok(())
     }
 
     #[instrument(skip(app_handle, event))]
-    pub fn on_system_tray_event(app_handle: &AppHandle, event: SystemTrayEvent) {
-        match event {
-            SystemTrayEvent::MenuItemClick { id, .. } => match id.as_str() {
-                mode @ ("rule_mode" | "global_mode" | "direct_mode" | "script_mode") => {
-                    let mode = &mode[0..mode.len() - 5];
-                    feat::change_clash_mode(mode.into());
-                }
+    pub fn on_menu_item_event(app_handle: &AppHandle, event: MenuEvent) {
+        let id = event.id().0.as_str();
+        match id {
+            mode @ ("rule_mode" | "global_mode" | "direct_mode" | "script_mode") => {
+                let mode = &mode[0..mode.len() - 5];
+                feat::change_clash_mode(mode.into());
+            }
 
-                "open_window" => resolve::create_window(app_handle),
-                "system_proxy" => feat::toggle_system_proxy(),
-                "tun_mode" => feat::toggle_tun_mode(),
-                "copy_env_sh" => feat::copy_clash_env("sh"),
-                #[cfg(target_os = "windows")]
-                "copy_env_cmd" => feat::copy_clash_env("cmd"),
-                #[cfg(target_os = "windows")]
-                "copy_env_ps" => feat::copy_clash_env("ps"),
-                "open_app_config_dir" => crate::log_err!(ipc::open_app_config_dir()),
-                "open_app_data_dir" => crate::log_err!(ipc::open_app_data_dir()),
-                "open_core_dir" => crate::log_err!(ipc::open_core_dir()),
-                "open_logs_dir" => crate::log_err!(ipc::open_logs_dir()),
-                "restart_clash" => feat::restart_clash_core(),
-                "restart_app" => help::restart_application(app_handle),
-                "quit" => {
-                    help::quit_application(app_handle);
-                }
-                _ => {
-                    proxies::on_system_tray_event(&id);
-                }
-            },
+            "open_window" => resolve::create_window(app_handle),
+            "system_proxy" => feat::toggle_system_proxy(),
+            "tun_mode" => feat::toggle_tun_mode(),
+            "copy_env_sh" => feat::copy_clash_env("sh"),
             #[cfg(target_os = "windows")]
-            SystemTrayEvent::LeftClick { .. } => {
-                resolve::create_window(app_handle);
+            "copy_env_cmd" => feat::copy_clash_env("cmd"),
+            #[cfg(target_os = "windows")]
+            "copy_env_ps" => feat::copy_clash_env("ps"),
+            "open_app_config_dir" => crate::log_err!(ipc::open_app_config_dir()),
+            "open_app_data_dir" => crate::log_err!(ipc::open_app_data_dir()),
+            "open_core_dir" => crate::log_err!(ipc::open_core_dir()),
+            "open_logs_dir" => crate::log_err!(ipc::open_logs_dir()),
+            "restart_clash" => feat::restart_clash_core(),
+            "restart_app" => help::restart_application(app_handle),
+            "quit" => {
+                help::quit_application(app_handle);
+            }
+            _ => {
+                proxies::on_system_tray_event(id);
+            }
+        }
+    }
+
+    pub fn on_system_tray_event(tray_icon: &TrayIcon, event: TrayIconEvent) {
+        match event {
+            TrayIconEvent::Click { button, .. } if button == MouseButton::Left => {
+                resolve::create_window(tray_icon.app_handle());
             }
             _ => {}
         }
