+++ backend/tauri/tauri.conf.json
@@ -1,118 +1,103 @@
 {
-  "package": {
-    "productName": "Clash Nyanpasu",
-    "version": "1.6.0"
-  },
-  "build": {
-    "distDir": "./.tmp/dist",
-    "devPath": "http://localhost:3000/",
-    "beforeDevCommand": "pnpm run web:dev",
-    "beforeBuildCommand": "pnpm run-p web:build generate:git-info && echo $(pwd) && rm -rf ./tauri/.tmp/dist && mv ../frontend/nyanpasu/dist ./tauri/.tmp/dist"
-  },
-  "tauri": {
-    "systemTray": {
-      "iconPath": "icons/win-tray-icon.png",
-      "iconAsTemplate": true
-    },
-    "bundle": {
-      "active": true,
-      "targets": "all",
-      "identifier": "moe.elaina.clash.nyanpasu",
-      "icon": [
-        "icons/32x32.png",
-        "icons/128x128.png",
-        "icons/128x128@2x.png",
-        "icons/icon.icns",
-        "icons/icon.ico"
-      ],
-      "resources": ["resources"],
-      "externalBin": [
-        "sidecar/clash",
-        "sidecar/mihomo",
-        "sidecar/mihomo-alpha",
-        "sidecar/clash-rs",
-        "sidecar/nyanpasu-service"
-      ],
-      "copyright": "© 2024 Clash Nyanpasu All Rights Reserved",
-      "category": "DeveloperTool",
-      "shortDescription": "Clash Nyanpasu! (∠・ω< )⌒☆",
-      "longDescription": "Clash Nyanpasu! (∠・ω< )⌒☆",
-      "deb": {
-        "depends": ["openssl"]
+  "bundle": {
+    "active": true,
+    "targets": "all",
+    "windows": {
+      "certificateThumbprint": null,
+      "digestAlgorithm": "sha256",
+      "timestampUrl": "",
+      "webviewInstallMode": {
+        "silent": true,
+        "type": "embedBootstrapper"
       },
-      "macOS": {
-        "frameworks": [],
-        "minimumSystemVersion": "11.0",
-        "exceptionDomain": "",
-        "signingIdentity": null,
-        "entitlements": null
+      "wix": {
+        "language": [
+          "zh-CN",
+          "en-US",
+          "ru-RU"
+        ],
+        "template": "./templates/installer.wxs",
+        "fragmentPaths": [
+          "./templates/cleanup.wxs"
+        ]
       },
-      "windows": {
-        "certificateThumbprint": null,
-        "digestAlgorithm": "sha256",
-        "timestampUrl": "",
-        "webviewInstallMode": {
-          "silent": true,
-          "type": "embedBootstrapper"
-        },
-        "wix": {
-          "language": ["zh-CN", "en-US", "ru-RU"],
-          "template": "./templates/installer.wxs",
-          "license": "../../LICENSE",
-          "fragmentPaths": ["./templates/cleanup.wxs"]
-        },
-        "nsis": {
-          "displayLanguageSelector": true,
-          "installerIcon": "icons/icon.ico",
-          "languages": ["SimpChinese", "English", "Russian"],
-          "license": "../../LICENSE",
-          "installMode": "both",
-          "template": "./templates/installer.nsi"
-        }
+      "nsis": {
+        "displayLanguageSelector": true,
+        "installerIcon": "icons/icon.ico",
+        "languages": [
+          "SimpChinese",
+          "English",
+          "Russian"
+        ],
+        "template": "./templates/installer.nsi",
+        "installMode": "both"
       }
     },
+    "icon": [
+      "icons/32x32.png",
+      "icons/128x128.png",
+      "icons/128x128@2x.png",
+      "icons/icon.icns",
+      "icons/icon.ico"
+    ],
+    "resources": [
+      "resources"
+    ],
+    "externalBin": [
+      "sidecar/clash",
+      "sidecar/mihomo",
+      "sidecar/mihomo-alpha",
+      "sidecar/clash-rs",
+      "sidecar/nyanpasu-service"
+    ],
+    "copyright": "© 2024 Clash Nyanpasu All Rights Reserved",
+    "category": "DeveloperTool",
+    "shortDescription": "Clash Nyanpasu! (∠・ω< )⌒☆",
+    "longDescription": "Clash Nyanpasu! (∠・ω< )⌒☆",
+    "macOS": {
+      "frameworks": [],
+      "minimumSystemVersion": "11.0",
+      "exceptionDomain": "",
+      "signingIdentity": null,
+      "entitlements": null
+    },
+    "linux": {
+      "deb": {
+        "depends": [
+          "openssl"
+        ]
+      }
+    },
+    "licenseFile": "../../LICENSE",
+    "createUpdaterArtifacts": "v1Compatible"
+  },
+  "build": {
+    "beforeBuildCommand": "pnpm run-p web:build generate:git-info && echo $(pwd) && rm -rf ./tauri/.tmp/dist && mv ../frontend/nyanpasu/dist ./tauri/.tmp/dist",
+    "frontendDist": "./.tmp/dist",
+    "beforeDevCommand": "pnpm run web:dev",
+    "devUrl": "http://localhost:3000/"
+  },
+  "productName": "Clash Nyanpasu",
+  "version": "1.6.0",
+  "identifier": "moe.elaina.clash.nyanpasu",
+  "plugins": {
     "updater": {
-      "active": true,
+      "pubkey": "dW50cnVzdGVkIGNvbW1lbnQ6IG1pbmlzaWduIHB1YmxpYyBrZXk6IDlBMUM0NjMxREZCNDRGMjYKUldRbVQ3VGZNVVljbW43N0FlWjA4UkNrbTgxSWxSSXJQcExXNkZjUTlTQkIyYkJzL0tsSWF2d0cK",
       "endpoints": [
         "https://mirror.ghproxy.com/https://github.com/LibNyanpasu/clash-nyanpasu/releases/download/updater/update-proxy.json",
         "https://gh-proxy.com/https://github.com/LibNyanpasu/clash-nyanpasu/releases/download/updater/update-proxy.json",
         "https://github.com/LibNyanpasu/clash-nyanpasu/releases/download/updater/update.json"
-      ],
-      "dialog": false,
-      "pubkey": "dW50cnVzdGVkIGNvbW1lbnQ6IG1pbmlzaWduIHB1YmxpYyBrZXk6IDlBMUM0NjMxREZCNDRGMjYKUldRbVQ3VGZNVVljbW43N0FlWjA4UkNrbTgxSWxSSXJQcExXNkZjUTlTQkIyYkJzL0tsSWF2d0cK"
-    },
-    "allowlist": {
-      "shell": {
-        "all": true
-      },
-      "window": {
-        "all": true
-      },
-      "process": {
-        "all": true
-      },
-      "globalShortcut": {
-        "all": true
-      },
-      "notification": {
-        "all": true
-      },
-      "os": {
-        "all": true
-      },
-      "clipboard": {
-        "all": true
-      },
-      "fs": {
-        "all": true
-      },
-      "dialog": {
-        "all": true
-      }
-    },
+      ]
+    }
+  },
+  "app": {
     "windows": [],
     "security": {
-      "csp": "script-src 'unsafe-eval' 'self'; default-src blob: data: filesystem: ws: wss: http: https: tauri: 'unsafe-eval' 'unsafe-inline' 'self'; img-src 'self' data: asset: blob: http://localhost:*  https:;"
+      "csp": "script-src 'unsafe-eval' 'self'; default-src blob: data: filesystem: ws: wss: http: https: tauri: 'unsafe-eval' 'unsafe-inline' 'self'; img-src 'self' data: asset: blob: http://localhost:*  https:;; connect-src ipc: http://ipc.localhost"
+    },
+    "trayIcon": {
+      "iconPath": "icons/win-tray-icon.png",
+      "iconAsTemplate": true
     }
   }
 }
