+++ frontend/nyanpasu/src/components/updater/updater-dialog.tsx
@@ -8,9 +8,9 @@ import { formatError } from "@/utils";
 import { message } from "@/utils/notification";
 import { Button } from "@mui/material";
 import { BaseDialog, BaseDialogProps, cn } from "@nyanpasu/ui";
-import { relaunch } from "@tauri-apps/api/process";
-import { open as openThat } from "@tauri-apps/api/shell";
-import { installUpdate, type UpdateManifest } from "@tauri-apps/api/updater";
+import { relaunch } from "@tauri-apps/plugin-process";
+import { open as openThat } from "@tauri-apps/plugin-shell";
+import { installUpdate, type UpdateManifest } from "@tauri-apps/plugin-updater";
 import styles from "./updater-dialog.module.scss";
 
 const Markdown = lazy(() => import("react-markdown"));
